Move frontier/provider filters to ScenarioExplorer; show all models on leaderboard
The leaderboard fits all 12 models comfortably as rows; it's the
scenario explorer's per-model column structure that actually gets wide.
Move the Frontier-only toggle and the provider chip group from the
leaderboard to the explorer, where they reduce horizontal density on
the prediction table.

The leaderboard now shows every model unconditionally (no filters, no
empty state, no "Clear filters" button). The explorer defaults to
Frontier only so the table opens at four columns; users can broaden by
toggling Frontier off or by clicking provider chips.

Also drop the "(analyst tolerance)" suffix from the Within-1% caption —
just say "Percent within 1% of reference."
